+++ packages/dashboard/src/components/DrivePeers.js
@@ -1,5 +1,7 @@
 import React from 'react'
 
+import { CopyToClipboard } from 'react-copy-to-clipboard'
+
 import { makeStyles } from '@material-ui/core'
 import Paper from '@material-ui/core/Paper'
 import Table from '@material-ui/core/Table'
@@ -8,6 +10,7 @@ import TableCell from '@material-ui/core/TableCell'
 import TableContainer from '@material-ui/core/TableContainer'
 import TableHead from '@material-ui/core/TableHead'
 import TableRow from '@material-ui/core/TableRow'
+import Tooltip from '@material-ui/core/Tooltip'
 
 import { humanizedBytes } from '../format'
 
@@ -32,7 +35,7 @@ function DrivePeers ({ peers }) {
       <Table aria-label="drive's peers table" size='small' className={classes.table}>
         <TableHead>
           <TableRow>
-            <TableCell>Peers</TableCell>
+            <TableCell><Tooltip title='Peers Ids' aria-label='peers ids column header' placement='right'><span>Peers</span></Tooltip></TableCell>
             <TableCell align='right'>Download Size</TableCell>
             <TableCell align='right'>Download Blocks</TableCell>
             <TableCell align='right'>Upload Size</TableCell>
@@ -43,7 +46,16 @@ function DrivePeers ({ peers }) {
           {peers.map(({ remoteAddress, downloadedBytes, downloadedBlocks, uploadedBytes, uploadedBlocks }) => {
             return (
               <TableRow key={remoteAddress}>
-                <TableCell className={classes.address}>{remoteAddress}</TableCell>
+                <TableCell className={classes.address}>
+                  <CopyToClipboard text={remoteAddress}>
+                    <code
+                      title={`${remoteAddress}\n(click to copy)`}
+                      className={classes.key}
+                    >
+                      {remoteAddress.substr(0, 6)}...{remoteAddress.substr(-6)}
+                    </code>
+                  </CopyToClipboard>
+                </TableCell>
                 <TableCell align='right'>{humanizedBytes(downloadedBytes).pretty}</TableCell>
                 <TableCell align='right'>{downloadedBlocks}</TableCell>
                 <TableCell align='right'>{humanizedBytes(uploadedBytes).pretty}</TableCell>